requirements.txt
apache-airflow-providers-microsoft-azure==3.0.0
